refactor(admin): ApplicationFormContextMenu -> ApplicationMenu로 이름 변경
컴포넌트명이 과도하게 길어 AdminPage 공통 컴포넌트명 관례에 맞게 단축
