+++ frontend/src/pages/AdminPage/components/ApplicationMenu/ApplicationMenu.styles.ts
