+++ frontend/src/pages/AdminPage/tabs/ApplicationTab/ApplicationListTab/components/mobile/ApplicationCardMobile/ApplicationCardMobile.tsx
@@ -1,6 +1,6 @@
 import type { MouseEvent, RefObject } from 'react';
 import MorebuttonIcon from '@/assets/images/icons/ellipsis_icon.svg?react';
-import ApplicationFormContextMenu from '@/pages/AdminPage/components/ApplicationFormContextMenu/ApplicationFormContextMenu';
+import ApplicationMenu from '@/pages/AdminPage/components/ApplicationMenu/ApplicationMenu';
 import {
   ApplicationFormItem,
   ApplicationFormStatus,
@@ -50,7 +50,7 @@ const ApplicationCardMobile = ({
           </Styled.MoreButton>
 
           {isMenuOpen && (
-            <ApplicationFormContextMenu
+            <ApplicationMenu
               isActive={isActive}
               onToggleStatus={() =>
                 onToggleStatus(application.id, application.status)
